+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.2
+- Add support Oracle data type.
+    - `CLOB`, `NCLOB`
+    - `NUMBER` with no length & scale specification
+- Miner fix.
+
 ## 1.1.1
 - Fix Postgres/Redshift parse of "::" syntax in field attribute.
 
@@ -7,8 +13,8 @@
 - Add `source_database` option.
 - Add `to_bigquery_fields` method to Columns dicttionary(`DdlParseColumnDict` class).
 - Fix BigQuery convert of Oracle data type.
-  - Oracle 'DATE' -> BigQuery 'DATETIME'
-  - Oracle 'NUMBER' -> BigQuery 'INTEGER' or 'FLOAT'
+    - Oracle 'DATE' -> BigQuery 'DATETIME'
+    - Oracle 'NUMBER' -> BigQuery 'INTEGER' or 'FLOAT'
 
 ## 1.0.2
 - Miner enhancement.